:truck: print_invite moved out of main.py
